requirements.txt
requests~=2.24.0
slackclient~=2.9.3
google-api-python-client~=1.12.5
google-auth-oauthlib~=0.4.1
